Completed Exercise 3
Successfully completed The Odin Projects exercise 3 (03-grouping-selectors). I initially created a separate class for the attributes that apply to both, and while it gave the intended result, it was not what the lesson wanted me to learn. So after reviewing the solution, I discovered the correct way to handle it was by using both unique classes separated by a comma. I can definitely see the benefit of doing it that way to cut down on clutter in the HTML file as the code gets more complicated down the road.