Completed Exercise 6 (Cascade Fix)
After much trial & error, and a few peeks at the solution file, I finally completed The Odin Project's exercise 6 (06-cascade-fix). In this lesson we were given a completed HTML file (not to touch) and a buggy CSS file. The task was to figure out what needed changed/removed/moved to get the HTML to function as intended. It was definitely the most challenging exercise so far.