Completed Exercise 02 - Flex Header
Completed the second flexbox exercise for the Odin Project. This one was definitely a major step up in difficulty and I finally had to peak at the solutions for help, but I was definitely on the right track. Just didn't realize you can put display: flex on a list too, but it makes sense now.