Corrected CSS
After reviewing the solution file, I discovered where I went wrong. The left and right containers just needed merged and the justify-content removed respectively. I then needed to add the proper margins, padding, and gap to the list. And voila, it's an exact match now. But I am proud of myself for getting it mostly right on my own. So that's a start.