Reviewed Exercise 5 - Modal
After the weekend, I struggled to get much of the exercise to work and needed a lot of help from the solution files. I got parts of the HTML and CSS correct, but overall, this was probably my toughest challenge yet. I plan to seek additional help on flexbox from other sources to supplement.